% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltas.R
\name{deltas}
\alias{deltas}
\alias{cru_clim}
\alias{deltas}
\title{CRU-based climatologies and delta change}
\usage{
cru_clim(x, limits = c(1960, 1989))

deltas(x, limits = c(1960, 1989))
}
\arguments{
\item{x}{a data frame such as returned using the \code{snapclim} package.}

\item{limits}{numeric, length-2 vector giving the range of years in the climatological window.}
}
\value{
a data frame.
}
\description{
Compute CRU-based monthly or seasonal climatologies using a specific historical window.
Use the climatologies to compute delta change from raw climate values.
}
\details{
\code{cru_clim} is used internally by \code{deltas}. It does not need to be used directly in this case.
}
\examples{
\dontshow{library(dplyr)}
library(snapclim)
library(dplyr)
x <- climdata("ar5stats", "Anchorage") \%>\%
  filter(Var \%in\% c("pr", "tas"))
cru_clim(x)
deltas(x)
}
