% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapstat.R
\docType{package}
\name{snapstat}
\alias{snapstat}
\alias{snapstat-package}
\title{snapstat: Graph and analyze SNAP data sets.}
\description{
Provides a suite of functions for graphing and analyzing SNAP data sets.
}
\details{
\code{snapstat} is the primary package in the SNAPverse for users to enagage in graphical and statistical analysis of SNAP data.
}
